benchmarks/EvoAgentBench/requirements.txt
pyyaml>=5.0
tomli>=1.1.0; python_version < "3.11"
httpx>=0.27.0

# SWE-bench
docker>=7.0.0
pandas>=2.0.0
swebench>=3.0.0

# BrowseComp-Plus: MCP search server
fastmcp>=3.0.0
faiss-cpu>=1.11.0.post1
datasets>=2.0.0,<3.0.0
transformers>=4.53.2
pydantic-settings>=2.0.0
torch>=2.4.0
torchvision>=0.19.0
accelerate>=1.9.0
peft>=0.16.0
tqdm>=4.67.1
python-dotenv>=1.1.1
pyserini>=1.2.0
pyngrok>=7.2.12
# flash-attn: install separately if GPU supports it
# pip install flash-attn or use pre-built wheel
qwen-omni-utils==0.0.8
tevatron @ git+https://github.com/texttron/tevatron.git@main

# LiveCodeBench: install separately (has conflicting deps)
# git clone https://github.com/LiveCodeBench/LiveCodeBench.git && pip install --no-deps -e ./LiveCodeBench
pebble>=5.1.0

# BrowseComp-Plus: LLM Judge
openai>=1.0.0

# GDPVal: evaluation (reading deliverable files)
python-docx>=0.8.10
openpyxl>=3.1.0
pdf2image>=1.16.0
Pillow>=9.0.0
